move single-use ArcsUtils.methods to where they are used
